Update PKHeX.Core NuGet
While the NuGet package is listed as 2024.01.13, I want to refer to it by the date it was released (2024.02.22) since it is more significant than just a bump for a bug fix.

+++ Unmaintained/PluginPile.Unmaintained.SwShRaidTool/Raid/RaidPKM.cs
@@ -1,9 +1,9 @@
 namespace PluginPile.Unmaintained.SwShRaidTool;
-public class RaidPKM(ushort species, byte altForm, uint ec, uint pid, int[] ivs, int ability, int abilityIdx, int gender, int nature, int deviation, uint shinyType, bool isGigantamax, sbyte forcedShinyType = 0) {
+public class RaidPKM(ushort species, byte altForm, uint ec, uint pid, int[] ivs, int ability, int abilityIdx, byte gender, int nature, int deviation, uint shinyType, bool isGigantamax, sbyte forcedShinyType = 0) {
   public readonly int Nature = nature;
   public readonly ushort Species = species;
   public readonly int[] IVs = ivs;
-  public readonly int Gender = gender;
+  public readonly byte Gender = gender;
   public readonly uint EC = ec;
   public readonly uint PID = pid;
   public readonly byte AltForm = altForm;
